Upd: refactor folder structure
