+++ server/middleware/authentication/Authenticator.ts
