+++ server/middleware/authentication/IAuthenticator.ts
